better separation of conscerns, module initializator takes repository object, add express app initializator
