--- src/modules/club/repository/clubs.repository.js
@@ -1,28 +1,0 @@
-module.exports = function makeClubsRepo(sequelizeModel) {
-	async function insert(clubInstance) {
-		const result = await sequelizeModel.create(clubInstance);
-		return result;
-	}
-
-	async function selectById(id) {
-		const found = sequelizeModel.findByPk(id);
-		return found;
-	}
-
-	async function selectByName(clubName) {
-		const found = sequelizeModel.findOne({ where: { clubName } });
-		return found;
-	}
-
-	async function selectAll() {
-		const found = sequelizeModel.findAll();
-		return found;
-	}
-
-	return Object.freeze({
-		insert,
-		selectAll,
-		selectById,
-		selectByName,
-	});
-};
