--- src/modules/club/service/find-club.service.js
@@ -1,6 +1,0 @@
-module.exports = function makeFindClub(clubRepo) {
-	return async function findClub(id) {
-		const result = await clubRepo.selectById(id);
-		return result;
-	};
-};
