--- src/modules/club/service/list-clubs.service.js
@@ -1,6 +1,0 @@
-module.exports = function makeListClubs(clubRepo) {
-	return async function listClubs() {
-		const result = await clubRepo.selectAll();
-		return result;
-	};
-};
